bug fixes done for acknowledgement screen

+++ apps/health_campaign_field_worker_app/lib/utils/utils.dart
@@ -580,7 +580,7 @@ DoseCriteriaModel? fetchProductVariant(
             .value
         : '0');
 
-    final weight = int.parse(individualModel.additionalFields != null &&
+    final weight = double.parse(individualModel.additionalFields != null &&
             individualModel.additionalFields!.fields
                 .where((element) => element.key == Constants.weight)
                 .isNotEmpty
